tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animationevent_init_after_dispatch.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - initAnimationEvent after dispatchEvent</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-events-">
   <meta name="flags" content="dom">
   <meta name="assert" content="initAnimationEvent may only be called before the
   AnimationEvent has been dispatched via the dispatchEvent method">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed;
         try {
            var evt = document.createEvent('AnimationEvent');
            evt.initAnimationEvent('animationend', true, false, 'Before', 0);
            div.dispatchEvent(evt);
            evt.initAnimationEvent('animationend', true, false, 'After', 1);
            passed = (evt.animationName == 'Before' && evt.elapsedTime == 0);
         } catch(e) {
            passed = false;
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
   <div></div>
</body>
</html>
